feat: 공구장 퇴장 시 state = -1, checkState 추가 feat : 새로운 기능 추가 fix : 버그 수정 docs : 문서 수정 modify : 기능 수정 및 코드 수정 refact : 리팩토링

+++ back/src/routers/groupRouter.js
@@ -52,6 +52,12 @@ groupRouter.put(
       const groupId = req.params.groupId;
       const quantity = req.body.quantity;
 
+      const checkState = await groupService.checkState({ groupId });
+      if (checkState === -1) {
+        const errorMessage = "가뭄이 들은 당근밭입니다.";
+        throw new Error(errorMessage);
+      }
+
       const updatedGroupInfo = await groupService.setQuantity({
         userId,
         groupId,
@@ -83,6 +89,12 @@ groupRouter.put(
       const userId = req.currentUserId;
       const groupId = req.params.groupId;
       const payment = req.body.payment;
+
+      const checkState = await groupService.checkState({ groupId });
+      if (checkState === -1) {
+        const errorMessage = "가뭄이 들은 당근밭입니다.";
+        throw new Error(errorMessage);
+      }
 
       const updatedGroupInfo = await groupService.setPayment({
         userId,
@@ -120,6 +132,12 @@ groupRouter.put(
       const deadline = req.body.deadline ?? null;
       const productId = req.body.productId ?? null;
       const state = req.body.state ?? null;
+
+      const checkState = await groupService.checkState({ groupId });
+      if (checkState === -1) {
+        const errorMessage = "가뭄이 들은 당근밭입니다.";
+        throw new Error(errorMessage);
+      }
 
       const toUpdate = {
         groupType,
@@ -251,6 +269,12 @@ groupRouter.put(
       const groupId = req.params.groupId;
       const { quantity, payment } = req.body;
 
+      const checkState = await groupService.checkState({ groupId });
+      if (checkState === -1) {
+        const errorMessage = "가뭄이 들은 당근밭입니다.";
+        throw new Error(errorMessage);
+      }
+
       const UpdatedGroup = await groupService.addParticipants({
         userId,
         groupId,
@@ -281,6 +305,12 @@ groupRouter.put(
     try {
       const userId = req.currentUserId;
       const groupId = req.params.groupId;
+
+      const checkState = await groupService.checkState({ groupId });
+      if (checkState === -1) {
+        const errorMessage = "가뭄이 들은 당근밭입니다.";
+        throw new Error(errorMessage);
+      }
 
       const UpdatedGroup = await groupService.deleteParticipant({
         userId,
